Complete task 1
